Fix incorrect matrix transformation, simplify.
The TBN matrix for conversion from tangent space to object space was
used to convert the light direction into tangent space, which is
incorrect. For this, the inverse must be used.

Now the blended normals are converted to object space, which simplifies
the lighting calculation. Also increases performance for multiple light
sources because only one transformation is needed in contrast to one per
light.