Fix issue with corrupted light lists.
For some reason the light lists as provided by
light_position_object_space gets corrupted when recreating the scene
manager. The end result is no working lighting when logging out and then
in again.
This fixes this by not using the Ogre provided
light_position_object_space auto constant function. Instead we're
recalculating the light position in object space per fragment. This is
of course less efficient, but will have to do for now.